requirements.txt
asn1crypto==0.24.0
azure-common==1.1.12
azure-nspkg==2.0.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
boto3==1.7.44
botocore==1.10.44
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.2.2
docutils==0.14
idna==2.7
jmespath==0.9.3
numpy==1.14.5
Pipewrench==0.2.4
pycparser==2.18
pycrypto==2.6.1
pymongo==3.6.1
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
termcolor==1.1.0
urllib3==1.23
